**Refactor context generation to zero-copy implementation, remove deprecated files** Switches to efficient zero-copy context generation, removes legacy Neo4j-based context generator, adds new scripts/docs.
