+++ orchestration/nana_orchestrator.py
@@ -40,7 +40,7 @@ from models import (
 )
 from models.user_input_models import UserStoryInputModel, user_story_to_objects
 from orchestration.chapter_flow import run_chapter_pipeline
-from processing.context_generator import generate_hybrid_chapter_context_native
+from processing.zero_copy_context_generator import generate_hybrid_chapter_context_native
 from processing.revision_logic import revise_chapter_draft_logic
 from processing.text_deduplicator import TextDeduplicator
 from ui.rich_display import RichDisplayManager
